--- a/README.md
+++ b/README.md
@@ -27,10 +27,10 @@ pip install -e .
 ```
 
 Some optional parts of the package require additional dependencies. 
-- Using commercial IP solvers to solve optimizations problems: `pip install qokit[solvers]`
-- GPU simulation: `pip install qokit[GPU]`
-- Development: `pip install qokit[dev]`
+- Using commercial IP solvers to solve optimizations problems: `pip install -e .[solvers]`
+- GPU simulation: `pip install -e .[GPU-CUDA12]`
 
+Please note that the GPU dependency is specified for CUDA 12x. For other versions of CUDA, please follow cupy installation instructions.
 
 If compilation fails, try installing just the Python version using `QOKIT_PYTHON_ONLY=1 pip install -e .`.