--- a/README.md
+++ b/README.md
@@ -27,8 +27,8 @@ pip install -e .
 ```
 
 Some optional parts of the package require additional dependencies. 
-- Using commercial IP solvers to solve optimizations problems: `pip install -e .[solvers]`
 - GPU simulation: `pip install -e .[GPU-CUDA12]`
+- Generating LP files to solve LABS using commercial IP solvers (`qokit/classical_methods`): `pip install -e .[solvers]`
 
 Please note that the GPU dependency is specified for CUDA 12x. For other versions of CUDA, please follow cupy installation instructions.