--- a/qokit/fur/nbcuda/fur.py
+++ b/qokit/fur/nbcuda/fur.py
@@ -5,43 +5,75 @@
 import math
 import numba.cuda
 import numpy as np
-
+import cupy as cp
+from pathlib import Path
+from functools import lru_cache
 
 ########################################
 # single-qubit X rotation
 ########################################
-@numba.cuda.jit
-def furx_kernel(x, wa, wb, q, mask1, mask2):
-    """CUDA kernel for fast uniform X rotations"""
-    n_states = len(x)
-    n_groups = n_states // 2
-    tid = numba.cuda.grid(1)
-
-    if tid < n_groups:
-        ia = (tid & mask1) | ((tid & mask2) << 1)
-        ib = ia | (1 << q)
-        x[ia], x[ib] = wa * x[ia] + wb * x[ib], wb * x[ia] + wa * x[ib]
-
-
-def furx(x: np.ndarray, theta: float, q: int):
+@lru_cache
+def get_furx_kernel(k_qubits: int, q_offset: int, state_mask: int):
     """
-    Apply in-place fast Rx gate exp(-1j * theta * X) on qubit q to statevector array x.
-    theta: rotation angle
+    Generate furx kernel for the specified sub-group size.
     """
-    n_states = len(x)
-    mask1 = (1 << q) - 1
-    mask2 = mask1 ^ ((n_states - 1) >> 1)
-    furx_kernel.forall(n_states)(x, math.cos(theta), -1j * math.sin(theta), q, mask1, mask2)
+    if k_qubits > 6:
+        kernel_name = f"furx_kernel<{k_qubits},{q_offset},{state_mask}>"
+    else:
+        kernel_name = f"warp_furx_kernel<{k_qubits},{q_offset}>"
+        
+    code = open(Path(__file__).parent / "furx.cu").read()
+    return cp.RawModule(code=code, name_expressions=[kernel_name],options=("-std=c++17",)).get_function(kernel_name)
 
+def furx(sv: cp.ndarray, a: float, b: float, k_qubits: int, q_offset: int, state_mask:int):
+    """
+    Apply in-place fast Rx gate exp(-1j * theta * X) on k consequtive qubits to statevector array x.
 
-def furx_all(x: np.ndarray, theta: float, n_qubits: int):
+    sv: statevector
+    a: cosine factor
+    b: sine factor
+    k_qubits: number of qubits to process concurrently
+    q_offset: starting qubit number
+    state_mask: mask for indexing
+    """
+    if k_qubits > 11:
+        raise ValueError("k_qubits should be <= 11 because of shared memory constraints")
+
+    seq_kernel = get_furx_kernel(k_qubits, q_offset, state_mask)
+
+    if k_qubits > 6:
+        threads = 1 << (k_qubits-1)
+    else:
+        threads = min(32, len(sv))
+
+    seq_kernel(
+            ((len(sv)//2+threads-1)//threads,),
+            (threads,),
+            (sv, a, b))    
+
+def furx_all(sv: np.ndarray, theta: float, n_qubits: int):
     """
     Apply in-place fast uniform Rx gates exp(-1j * theta * X) to statevector array x.
-    theta: rotation angle
-    """
-    for i in range(n_qubits):
-        furx(x, theta, i)
 
+    sv: statevector
+    theta: rotation angle
+    n_qubits: total number of qubits
+    """
+    n_states = len(sv)
+    state_mask = ((n_states - 1) >> 1)
+
+    a,b = math.cos(theta), -math.sin(theta)   
+
+    group_size = 10
+    last_group_size = n_qubits%group_size
+
+    cp_sv = cp.asarray(sv)
+
+    for q_offset in range(0, n_qubits-last_group_size, group_size):
+       furx(cp_sv, a, b, group_size, q_offset, state_mask)
+
+    if last_group_size > 0:
+        furx(cp_sv, a, b, last_group_size, n_qubits-last_group_size, state_mask)
 
 ########################################
 # two-qubit XX+YY rotation