## Overview
We added 2 new CUDA kernels to accelerate x-rotations. They increase the speed of each rotation by ~2.5x, based on this [benchmark](https://github.com/mebenstein-vorticity/QOKit/blob/performance/examples/benchmark_example.ipynb). 
We got the following results on a NVIDIA T4:

| **N Qubits** | **LABS (Before)** | **MaxCut (Before)** | **LABS (After)** | **MaxCut (After)** | **Improvement** |
|--------------|-------------------|---------------------|------------------|--------------------|-----------------|
|    **22**    |       1.5177      |        1.5227       |      0.5929      |       0.5926       |      2.56x      |
|    **24**    |       6.6131      |        6.6295       |      2.4754      |       2.4794       |      2.67x      |
|    **26**    |      28.5703      |       28.5143       |      11.5100     |       11.4985      |      2.48x      |

## Approach
The in-place rotation kernel is applied $n$ times for $n$ qubits to perform a single rotation. We found that for every $k < n$ steps in the operation, we can divide the state vector into $2^{n-k}$ independent sub-groups. This allows us to process groups of states independently (requiring fewer synchronizations) and ensure that for each group all memory accesses are perfectly cached. 

The `furx_kernel` uses shared memory to store all values for a given sub-group. $k$ is limited by the maximum shared memory size of the GPU and the maximum number of threads per block. For most modern GPUs we can at most allocate 64KB of shared memory per block, which is slightly below $16*2^{12}$ and thus limits us to $k \le 11$. We find that we achieve the best performance for $k = 10$, most likely because that requires only $512$ threads and achieves higher occupancy. We tried partitioning the data such that each thread is responsible for multiple rows, which gives us control over the number of threads needed independently of the shared memory requirement. Although this can lead to higher occupancy, we saw an overall decrease in performance and thus removed the partitioning. This mean $k$ is additionally limited to $\le 11$ because the kernel requires $t = 2^{k-1}$ threads per block and cuda only support a maximum of $t = 1024$.

The `warp_furx_kernel` does not use shared memory, but instead uses only registers and [warp-shuffle](https://docs.nvidia.com/cuda/cuda-c-programming-guide/index.html#warp-shuffle-functions) primitives to exchange data. This implies the kernel is limited to $k \le 6$ because currently each warp has 32 threads. For $k \lt 6$ each warp handles multiple groups to ensure peak occupancy. 

To perform a rotation we split up the $n$ steps into a sequence of calls to the `furx_kernel` with $k=10$ and a final call if $n$ is not a multiple of 10 to either the `furx_kernel` if $n \mod 10 \gt 6$ or the `warp_furx_kernel` otherwise. In total, instead of $n$ kernel calls, we know have $\lceil n/10 \rceil$ kernel calls for each rotation.

## Possible Future Improvements
For the  `warp_furx_kernel` we only have 4 global memory accesses per thread, and no other memory accesses. The `furx_kernel` also only has 4 global memory accesses, but has $4*k$ shared memory accesses. It is possible to incorporate the warp level exchange into this kernel and to handle it as a sequence of 2 sub-groups. For example if $k=10$, we could process this as one $k=6$ and one $k=4$ (or whatever the most efficient distribution is) sub-group, where each sub-group can work purely on the warp level. This would reduce the shared memory accesses to $8$ for each thread, and only require $2$ block-level synchronizations, instead of $k$. 

Furthermore, in terms of memory requirements, this approach of processing groups of states independently can allow us to work with a large number of qubits without requiring the whole state vector to be in memory at the same time. Groups could either be processes on multiple GPUs concurrently, or in sequence on a single GPU.