@rsln-s 2024-07-24
Confirmed the claimed performance:

```
	LABS finished in 0.5839 on average, min: 0.5792, max: 0.5895
	MaxCut finished in 0.5835 on average, min: 0.5790, max: 0.5872
N=24
	LABS finished in 2.4353 on average, min: 2.4193, max: 2.4506
	MaxCut finished in 2.4457 on average, min: 2.4352, max: 2.4561
N=26
	LABS finished in 11.3791 on average, min: 11.2953, max: 11.4151
	MaxCut finished in 11.3921 on average, min: 11.3866, max: 11.4027
```