@mebenstein-vorticity 2024-07-24
> @mebenstein-vorticity I'm seeing that one of the tests breaks after the change to cupy. Here's a full trace:
> 
> ```
> $ pytest -s -k test_maxcut_qaoa_obj_consistency_across_simulators
> ==================================================== test session starts =====================================================
> platform linux -- Python 3.9.4, pytest-8.3.1, pluggy-1.5.0
> rootdir: /opt/omniai/work/instance1/jupyter/QOKit
> configfile: pyproject.toml
> collecting 1 item                                                                                                            [<class 'qokit.fur.nbcuda.qaoa_simulator.QAOAFURXSimulatorGPU'>, <class 'qokit.fur.python.qaoa_simulator.QAOAFURXSimulator'>]
> collected 113 items / 112 deselected / 1 selected                                                                            
> 
> tests/test_qaoa_objective_maxcut.py F
> 
> ========================================================== FAILURES ==========================================================
> _____________________________________ test_maxcut_qaoa_obj_consistency_across_simulators _____________________________________
> 
>     def test_maxcut_qaoa_obj_consistency_across_simulators():
>         N = 8
>         for d, p in [(3, 11), (5, 4)]:
>             G = nx.random_regular_graph(d, N)
>             gamma, beta = get_fixed_gamma_beta(d, p)
>             for objective in ["expectation", "overlap"]:
>                 qaoa_objectives = [
>                     get_qaoa_maxcut_objective(N, p, G=G, parameterization="gamma beta", simulator=simulator, objective=objective)(gamma, beta)
>                     for simulator in simulators_to_run_names
>                 ]
> >               assert np.all(np.isclose(qaoa_objectives, qaoa_objectives[0]))
> 
> tests/test_qaoa_objective_maxcut.py:80: 
> _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ 
> <__array_function__ internals>:180: in isclose
>     ???
> cupy/_core/core.pyx:1717: in cupy._core.core._ndarray_base.__array_function__
>     ???
> ../py39/lib/python3.9/site-packages/cupy/_logic/comparison.py:124: in isclose
>     a = cupy.asanyarray(a)
> ../py39/lib/python3.9/site-packages/cupy/_creation/from_data.py:102: in asanyarray
>     return _core.array(a, dtype, False, order, blocking=blocking)
> cupy/_core/core.pyx:2383: in cupy._core.core.array
>     ???
> cupy/_core/core.pyx:2410: in cupy._core.core.array
>     ???
> cupy/_core/core.pyx:2546: in cupy._core.core._array_default
>     ???
> _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ 
> 
> >   ???
> E   TypeError: Implicit conversion to a NumPy array is not allowed. Please use `.get()` to construct a NumPy array explicitly.
> 
> cupy/_core/core.pyx:1479: TypeError
> ```
> 
> Examining `qaoa_objectives` suggests that the error is caused by the GPU objective for "overlap" now returning a cupy array instead of a float:
> 
> ```
> In [1]: [type(x) for x in qaoa_objectives]
> Out[1]: [cupy.ndarray, numpy.float64, numpy.float64]
> 
> In [2]: qaoa_objectives
> Out[2]: [array(0.13751705), 0.13751704613944205, 0.1375170461394406]
> ```
> 
> Could you please fix? In general, it would be appreciated if you could run the tests locally before opening a PR. Thanks!

I ran the unit tests before and after my changes, and this test was already failing before. You can see in my changes that we never change the array type and never return a cupy array. I only create a wrapper around the numba memory pointer and pass it to the the functions internally. I will look into it, but I am quite sure this error was there before.