@mebenstein-vorticity 2024-07-24
I found the issue. Like I said, it was there before my changes. In the `get_overlap` method, it is trying to import cupy (which failed before because it was not in the dependencies). Now that it is included the test failed. I solved it by calling `.item()` on the output return value.